Change amounts type to decimal

+++ src/campaigns/models.py
@@ -2,7 +2,7 @@ import enum
 import uuid
 from datetime import datetime
 from typing import TYPE_CHECKING
-
+from decimal import Decimal
 from sqlalchemy import DateTime, Enum as SAEnum, ForeignKey, Numeric, String, Text
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import Mapped, mapped_column, relationship
@@ -34,8 +34,8 @@ class Campaign(Base):
     title: Mapped[str] = mapped_column(String(200), nullable=False, index=True)
     description: Mapped[str] = mapped_column(String(500), nullable=False)
     story: Mapped[str | None] = mapped_column(Text, nullable=True)
-    goal_amount: Mapped[float] = mapped_column(Numeric(10, 2), nullable=False)
-    current_amount: Mapped[float] = mapped_column(Numeric(10, 2), default=0.00)
+    goal_amount: Mapped[Decimal] = mapped_column(Numeric(10, 2), nullable=False)
+    current_amount: Mapped[Decimal] = mapped_column(Numeric(10, 2), default=0.00)
     image_url: Mapped[str | None] = mapped_column(String, nullable=True)
     deadline: Mapped[datetime] = mapped_column(DateTime(timezone=True), nullable=False)
     status: Mapped[CampaignStatus] = mapped_column(
